refactor: extracted DataValue validation to validator methods [DHIS2-6163]
